增加异常处理

+++ src/main/java/cn/godbol/common/service/api/DefaultDeleteService.java
@@ -1,5 +1,7 @@
 package cn.godbol.common.service.api;
 
+import cn.godbol.common.service.exception.EntityNotFoundException;
+import cn.godbol.common.service.exception.ServerErrorConstant;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 import java.io.Serializable;
@@ -12,7 +14,12 @@ public interface DefaultDeleteService<T, ID extends Serializable> extends Delete
 
     @Override
     default void delete(ID id){
-        getRepository().delete(id);
+        if (getRepository().exists(id)){
+            getRepository().delete(id);
+        } else {
+            throw new EntityNotFoundException(this.getClass().getName(), ServerErrorConstant.DELETE, getEntityName(), id.toString());
+        }
+
     }
 
     @Override
